Fin de la aplicacion
Aplicacion de Blog Terminada, basica

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,20 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -42,11 +52,25 @@ class PostsController extends Controller
         $this->validate($request,
         array(
             'title' =>'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' =>'image|nullable|max:1999'
         ));
+        //manejador de archivo
+        if($request->hasFile('cover_image')){
+            //obtener el nombre y la ext
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $filename= pathinfo($filenameWithExt,PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
         $post = new Post();
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
          return redirect('/posts')->with('success','Se a Publicado Su Post');
     }
@@ -72,6 +96,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Sin Autorizacion');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -89,11 +116,22 @@ class PostsController extends Controller
             'title' =>'required',
             'body' => 'required'
         ));
+        if($request->hasFile('cover_image')){
+            //obtener el nombre y la ext
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $filename= pathinfo($filenameWithExt,PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
         $post =  Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
-         return redirect('/posts')->with('success','Se actualizo  su Post');
+         return redirect('/dashboard')->with('success','Se actualizo  su Post');
     }
 
     /**
@@ -105,7 +143,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user->id){
+            return redirect('/posts')->with('error','Sin Autorizacion');
+        }
+        if($post->cover_image != 'noimage.jpg'){
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
-        return redirect('/posts')->with('success','Se elimino la publicacion');
+        return redirect('/dashboard')->with('success','Se elimino la publicacion');
     }
 }
